feat(validation): convert to an owned builder-type pattern for Validator
This approach facilitates a form of chaining that let's us do more
convenient assertions and early Result returns in constructor methods.
It would be nice to also accumulate the variables inside the Validator
but this is not trivial if we want it all in memory rather than using
something heap allocated or added complexity.