feat: update dependencies to latest versions
One thing to note is that Poisson in rand_distr now returns a floating
point value. Upon inspection, the team working on rand decided that this
makes for a cleaner interface, and thus ask that users cast the floats
to integers. This isn't the most beautiful approach, but it should be
sufficient.